requirements.txt
Django==4.2.7
dj-database-url==2.1.0
django-haystack==3.2.1
psycopg2-binary==2.9.5
gunicorn==21.2.0
djangorestframework==3.14.0
pillow==10.2.0